+++ components/navbar/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+// import Link from 'next/link'
+
+export default function NavBar () {
+  return (
+    <nav>
+      {/* <Link>
+        TEst
+      </Link> */}
+    </nav>
+  )
+}
