+++ config/chartjs.config.js
@@ -1,0 +1,10 @@
+export const CommonOptionsChart = {
+  interaction: {
+    mode: 'index',
+    axis: 'y'
+  },
+  plugins: {
+    legend: { position: 'bottom' },
+    tooltips: { position: 'nearest' }
+  }
+}
